sales-pipeline/airflow/requirements.txt
apache-airflow-providers-postgres
astronomer-cosmos
dbt-postgres